refactor: split commands to separate classes
