+++ src/main/java/dev/dfonline/codeclient/command/impl/CommandWorldPlot.java
@@ -1,0 +1,36 @@
+package dev.dfonline.codeclient.command.impl;
+
+import com.mojang.brigadier.builder.LiteralArgumentBuilder;
+import dev.dfonline.codeclient.CodeClient;
+import dev.dfonline.codeclient.command.Command;
+import dev.dfonline.codeclient.location.Plot;
+import net.fabricmc.fabric.api.client.command.v2.FabricClientCommandSource;
+
+import static net.fabricmc.fabric.api.client.command.v2.ClientCommandManager.literal;
+
+public class CommandWorldPlot extends Command {
+    @Override
+    public String name() {
+        return "worldplot";
+    }
+
+    @Override
+    public LiteralArgumentBuilder<FabricClientCommandSource> create(LiteralArgumentBuilder<FabricClientCommandSource> cmd) {
+        return cmd.executes(context -> {
+            if (CodeClient.location instanceof Plot plot) plot.setSize(null);
+            return 0;
+        }).then(literal("basic").executes(context -> {
+            if (CodeClient.location instanceof Plot plot) plot.setSize(Plot.Size.BASIC);
+            return 0;
+        })).then(literal("large").executes(context -> {
+            if (CodeClient.location instanceof Plot plot) plot.setSize(Plot.Size.LARGE);
+            return 0;
+        })).then(literal("massive").executes(context -> {
+            if (CodeClient.location instanceof Plot plot) plot.setSize(Plot.Size.MASSIVE);
+            return 0;
+        })).then(literal("mega").executes(context -> {
+            if (CodeClient.location instanceof Plot plot) plot.setSize(Plot.Size.MEGA);
+            return 0;
+        }));
+    }
+}
